% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minipkg-package.R
\docType{package}
\name{minipkg-package}
\alias{minipkg}
\alias{minipkg-package}
\title{minipkg: Mini Package for Reproducible Examples}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Florian Privé \email{florian.prive.21@gmail.com}

}
\keyword{internal}
