% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-parallel.R
\name{myfunpar}
\alias{myfunpar}
\title{Use foreach}
\usage{
myfunpar(x, ncores)
}
\arguments{
\item{x}{A vector.}

\item{ncores}{Number of cores to use.}
}
\value{
The \code{x} vector, squared.
}
\description{
Use foreach
}
\examples{
myfunpar(1:3, 2)
}
