% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.R
\name{what_time}
\alias{what_time}
\title{Indicate what time it is}
\usage{
what_time(language = "fr")
}
\arguments{
\item{language}{string indicating the language}
}
\value{
A string with a sentence
}
\description{
Here we would put more details
}
\examples{
what_time()
what_time(language = "en")
}
