% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/current-time.R
\name{what_time}
\alias{what_time}
\title{Return current time}
\usage{
what_time(language = "fr")
}
\arguments{
\item{language}{language to use (either "en" or "fr")}
}
\value{
A character string
}
\description{
Return current time
}
\examples{
what_time()
}
