% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.R
\name{what_time}
\alias{what_time}
\title{Time function}
\usage{
what_time(language = "es")
}
\arguments{
\item{language}{Language to use, "es" o "en".  Default is "es"}
}
\value{
A string with the current time.
}
\description{
Time function
}
\examples{
what_time()
what_time("en")
}
